Using “tgmath.h” so that the correct math functions are automatically used for CGFloat.

+++ Source/WaveChannelView.m
@@ -24,6 +24,8 @@
 #import "WaveChannelView.h"
 #import "WaveRepresentationView.h"
 
+#import <tgmath.h>
+
 
 @interface WaveChannelView () <WaveRepresentationDelegate>
 @end
@@ -66,7 +68,7 @@
 {
     if (!_sampleArray) return;
 
-    NSInteger newTileCount = pow(2, floor(log2(ceil(_magnification))));
+    NSInteger newTileCount = pow((CGFloat)2, floor(log2(ceil(_magnification))));
 
     if ([_activeRepresentation tileCount] != newTileCount) {
         NSInteger sampleCount = [[self enclosingScrollView] frame].size.width * newTileCount;
